requirements.txt
 
flask
flask-cors
kafka-python
pandas
plotly
